LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-extensions-header-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-extensions-header' failed: Received unexpected extension: x-foo
Test whether WebSocket handshake fails if the server sends Sec-WebSocket-Extensions header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
